Add unified BitBox02 CLI
Introduce bitbox02_cli.py as the canonical Python CLI with explicit interactive mode and non-interactive subcommands for device, wallet, bootloader, and firmware workflows.

Keep send_message.py and load_firmware.py as compatibility wrappers, and update docs plus Python linting to point at the new entrypoint.
