+++ scripts/lint-python
@@ -23,7 +23,7 @@ ARGS=($(find py releases -name '*.py' | grep -v -e generated -e old))
 # behavior).
 # We must typecheck the whole `py` directory even if only a few files are modified so that mypy
 # sees all types
-${MYPY} --implicit-reexport --namespace-packages --ignore-missing-imports --strict py/send_message.py py/bitbox02/bitbox02 releases/describe_signed_firmware.py
+${MYPY} --implicit-reexport --namespace-packages --ignore-missing-imports --strict py/bitbox02_cli.py py/bitbox02/bitbox02 releases/describe_signed_firmware.py
 
 # Must run from root directory where .pylintrc is
 # We ignore refactor and convention messages because they can differ with `black`
